This keeps branch in sync with trunk.  At the moment only basic GPU functionality works.  No contiguous arrays approach, RMCRT doesn't work here,  Wasatch needs to be updated, etc.

+++ src/CCA/Components/Models/Radiation/RMCRT/RMCRTCommon.cc
@@ -37,7 +37,7 @@
 
 #include <include/sci_defs/uintah_testdefs.h.in>
 
-
+#define DEBUG -9 // 1: divQ, 2: boundFlux, 3: scattering
 //______________________________________________________________________
 //
 using namespace Uintah;
@@ -58,6 +58,8 @@ bool   RMCRTCommon::d_isSeedRandom;
 bool   RMCRTCommon::d_allowReflect;
 int    RMCRTCommon::d_matl;
 string RMCRTCommon::d_abskgBC_tag;
+vector<IntVector> RMCRTCommon::d_dbgCells;
+
 
 MaterialSet* RMCRTCommon::d_matlSet = 0;
 const VarLabel* RMCRTCommon::d_sigmaT4Label;
@@ -85,6 +87,7 @@ RMCRTCommon::RMCRTCommon( TypeDescription::Type FLT_DBL )
 
   d_gac     = Ghost::AroundCells;
   d_gn      = Ghost::None;
+  d_flowCell = -1; //<----HARD CODED FLOW CELL
 }
 
 //______________________________________________________________________
@@ -94,7 +97,9 @@ RMCRTCommon::RMCRTCommon( TypeDescription::Type FLT_DBL )
 RMCRTCommon::~RMCRTCommon()
 {
   VarLabel::destroy( d_sigmaT4Label );
-  VarLabel::destroy( d_abskgLabel );
+  if (RMCRTCommon::d_FLT_DBL == TypeDescription::float_type){
+    VarLabel::destroy( d_abskgLabel );
+  }
 
   // when the radiometer class is invoked d_matlSet it deleted twice.  This prevents that.
   if( d_matlSet ) {
@@ -121,8 +126,8 @@ RMCRTCommon::registerVarLabels(int   matlIndex,
   d_cellTypeLabel   = celltype;
   d_divQLabel       = divQ;
 
-  d_abskgBC_tag = d_compAbskgLabel->getName(); // The label name changes when using floats. 
-  
+  d_abskgBC_tag = d_compAbskgLabel->getName(); // The label name changes when using floats.
+
   // If using RMCRT:DBL
   const Uintah::TypeDescription* td = d_compAbskgLabel->typeDescription();
   const Uintah::TypeDescription::Type subtype = td->getSubType()->getType();
@@ -144,9 +149,9 @@ RMCRTCommon::registerVarLabels(int   matlIndex,
 
 //______________________________________________________________________
 //  This task will convert the CCVariable abskg from double -> float
-//  If abskg is of type double and the component has                  
-//  specified that RMCRT communicate the all-to-all variables (abskg & sigmaT4)  
-//  as a float then convert abskg to float                      
+//  If abskg is of type double and the component has
+//  specified that RMCRT communicate the all-to-all variables (abskg & sigmaT4)
+//  as a float then convert abskg to float
 //______________________________________________________________________
 void
 RMCRTCommon::sched_DoubleToFloat( const LevelP& level,
@@ -197,9 +202,9 @@ RMCRTCommon::DoubleToFloat( const ProcessorGroup*,
   for (int p=0; p < patches->size(); p++){
     const Patch* patch = patches->get(p);
     printTask(patches,patch,dbg,"Doing RMCRTCommon::DoubleToFloat");
-    
+
     constCCVariable<double> abskg_D;
-    CCVariable< float > abskg_F; 
+    CCVariable< float > abskg_F;
 
     DataWarehouse* myDW = new_dw->getOtherDataWarehouse(which_dw);
     myDW->get(abskg_D,             d_compAbskgLabel, d_matl, patch, Ghost::None, 0);
@@ -349,8 +354,8 @@ RMCRTCommon::rayLocation( MTRand& mTwister,
                          Vector& location)
 {
   if( useCCRays == false ){
-    location[0] =   origin[0] +  mTwister.rand() ;
-    location[1] =   origin[1] +  mTwister.rand() * DyDx ;
+    location[0] =   origin[0] +  mTwister.rand() ;             // FIX ME!!! This is not the physical location of the ray.
+    location[1] =   origin[1] +  mTwister.rand() * DyDx ;      // this is index space.
     location[2] =   origin[2] +  mTwister.rand() * DzDx ;
   }else{
     location[0] =   origin[0] +  0.5 ;
@@ -386,7 +391,7 @@ RMCRTCommon::reflect(double& fs,
 }
 
 //______________________________________________________________________
-//    Core function:  Integrate the intensity
+//    Integrate the intensity
 //______________________________________________________________________
 template <class T >
 void
@@ -402,196 +407,200 @@ RMCRTCommon::updateSumI ( Vector& ray_direction,
                          MTRand& mTwister)
 
 {
-/*`==========TESTING==========*/
-#if DEBUG == 1
-  printf("        updateSumI: [%d,%d,%d] ray_dir [%g,%g,%g] ray_loc [%g,%g,%g]\n", origin.x(), origin.y(), origin.z(),ray_direction.x(), ray_direction.y(), ray_direction.z(), ray_location.x(), ray_location.y(), ray_location.z());
-#endif
-/*===========TESTING==========`*/
+
 
   IntVector cur = origin;
   IntVector prevCell = cur;
   // Step and sign for ray marching
-   int step[3];                                          // Gives +1 or -1 based on sign
-   bool sign[3];
+  int step[3];                                          // Gives +1 or -1 based on sign
+  bool sign[3];
+  Vector inv_ray_direction = Vector(1.0)/ray_direction;
 
-   Vector inv_ray_direction = Vector(1.0)/ray_direction;
-   findStepSize(step, sign, inv_ray_direction);
-   Vector D_DxRatio(1, Dx.y()/Dx.x(), Dx.z()/Dx.x() );
+/*`==========TESTING==========*/
+#if DEBUG == 1
+  if( isDbgCell(origin) ) {
+    printf("        updateSumI: [%d,%d,%d] ray_dir [%g,%g,%g] ray_loc [%g,%g,%g]\n", origin.x(), origin.y(), origin.z(),ray_direction.x(), ray_direction.y(), ray_direction.z(), ray_location.x(), ray_location.y(), ray_location.z());
+    printf("        inv_ray_dir [%g,%g,%g]\n", inv_ray_direction.x(), inv_ray_direction.y(), inv_ray_direction.z());
+  }
+#endif
+/*===========TESTING==========`*/
 
-   Vector tMax;         // (mixing bools, ints and doubles)
-   tMax.x( (origin[0] + sign[0]                - ray_location[0]) * inv_ray_direction[0] );
-   tMax.y( (origin[1] + sign[1] * D_DxRatio[1] - ray_location[1]) * inv_ray_direction[1] );
-   tMax.z( (origin[2] + sign[2] * D_DxRatio[2] - ray_location[2]) * inv_ray_direction[2] );
+  findStepSize(step, sign, inv_ray_direction);
+  Vector D_DxRatio(1, Dx.y()/Dx.x(), Dx.z()/Dx.x() );
 
-   //Length of t to traverse one cell
-   Vector tDelta = Abs(inv_ray_direction) * D_DxRatio;
+  Vector tMax;         // (mixing bools, ints and doubles)
+  tMax.x( (origin[0] + sign[0]                - ray_location[0]) * inv_ray_direction[0] );
+  tMax.y( (origin[1] + sign[1] * D_DxRatio[1] - ray_location[1]) * inv_ray_direction[1] );
+  tMax.z( (origin[2] + sign[2] * D_DxRatio[2] - ray_location[2]) * inv_ray_direction[2] );
 
-   //Initializes the following values for each ray
-   bool in_domain     = true;
-   double tMax_prev   = 0;
-   double intensity   = 1.0;
-   double fs          = 1.0;
-   int nReflect       = 0;                 // Number of reflections
-   double optical_thickness      = 0;
-   double expOpticalThick_prev   = 1.0;
+  //Length of t to traverse one cell
+  Vector tDelta = Abs(inv_ray_direction) * D_DxRatio;
+
+  //Initializes the following values for each ray
+  bool in_domain     = true;
+  double tMax_prev   = 0;
+  double intensity   = 1.0;
+  double fs          = 1.0;
+  int nReflect       = 0;                 // Number of reflections
+  double optical_thickness      = 0;
+  double expOpticalThick_prev   = 1.0;
 
 
 #ifdef RAY_SCATTER
-   double scatCoeff = d_sigmaScat;          //[m^-1]  !! HACK !! This needs to come from data warehouse
-   if (scatCoeff == 0) scatCoeff = 1e-99;  // avoid division by zero
+  double scatCoeff = d_sigmaScat;          //[m^-1]  !! HACK !! This needs to come from data warehouse
+  if (scatCoeff == 0) scatCoeff = 1e-99;  // avoid division by zero
 
-   // Determine the length at which scattering will occur
-   // See CCA/Components/Arches/RMCRT/PaulasAttic/MCRT/ArchesRMCRT/ray.cc
-   double scatLength = -log(mTwister.randDblExc() ) / scatCoeff;
-   double curLength = 0;
+  // Determine the length at which scattering will occur
+  // See CCA/Components/Arches/RMCRT/PaulasAttic/MCRT/ArchesRMCRT/ray.cc
+  double scatLength = -log(mTwister.randDblExc() ) / scatCoeff;
+  double curLength = 0;
 #endif
 
-   //+++++++Begin ray tracing+++++++++++++++++++
-   //Threshold while loop
-   while ( intensity > d_threshold ){
+  //+++++++Begin ray tracing+++++++++++++++++++
+  //Threshold while loop
+  while ( intensity > d_threshold ){
 
-     DIR face = NONE;
+    DIR dir = NONE;
 
-     while (in_domain){
+    while (in_domain){
 
-       prevCell = cur;
-       double disMin = -9;          // Represents ray segment length.
+      prevCell = cur;
+      double disMin = -9;          // Represents ray segment length.
 
-       T abskg_prev = abskg[prevCell];  // optimization
-       T sigmaT4OverPi_prev = sigmaT4OverPi[prevCell];
-       //__________________________________
-       //  Determine which cell the ray will enter next
-       if ( tMax[0] < tMax[1] ){        // X < Y
-         if ( tMax[0] < tMax[2] ){      // X < Z
-           face = X;
-         } else {
-           face = Z;
-         }
-       } else {
-         if( tMax[1] < tMax[2] ){       // Y < Z
-           face = Y;
-         } else {
-           face = Z;
-         }
-       }
+      T abskg_prev = abskg[prevCell];  // optimization
+      T sigmaT4OverPi_prev = sigmaT4OverPi[prevCell];
+      //__________________________________
+      //  Determine which cell the ray will enter next
+      if ( tMax[0] < tMax[1] ){        // X < Y
+        if ( tMax[0] < tMax[2] ){      // X < Z
+          dir = X;
+        } else {
+          dir = Z;
+        }
+      } else {
+        if( tMax[1] < tMax[2] ){       // Y < Z
+          dir = Y;
+        } else {
+          dir = Z;
+        }
+      }
 
-       //__________________________________
-       //  update marching variables
-       cur[face]  = cur[face] + step[face];
-       disMin     = (tMax[face] - tMax_prev);
-       tMax_prev  = tMax[face];
-       tMax[face] = tMax[face] + tDelta[face];
+      //__________________________________
+      //  update marching variables
+      cur[dir]  = cur[dir] + step[dir];
+      disMin     = (tMax[dir] - tMax_prev);
+      tMax_prev  = tMax[dir];
+      tMax[dir] = tMax[dir] + tDelta[dir];
 
-       ray_location[0] = ray_location[0] + (disMin  * ray_direction[0]);
-       ray_location[1] = ray_location[1] + (disMin  * ray_direction[1]);
-       ray_location[2] = ray_location[2] + (disMin  * ray_direction[2]);
+      ray_location[0] = ray_location[0] + (disMin  * ray_direction[0]);
+      ray_location[1] = ray_location[1] + (disMin  * ray_direction[1]);
+      ray_location[2] = ray_location[2] + (disMin  * ray_direction[2]);
 
- /*`==========TESTING==========*/
+/*`==========TESTING==========*/
 #if DEBUG == 1
-if(origin.x() == 0 && origin.y() == 0 && origin.z() ==0){
-    printf( "            cur [%d,%d,%d] prev [%d,%d,%d] ", cur.x(), cur.y(), cur.z(), prevCell.x(), prevCell.y(), prevCell.z());
-    printf( " face %d ", face );
-    printf( "tMax [%g,%g,%g] ",tMax.x(),tMax.y(), tMax.z());
-    printf( "rayLoc [%g,%g,%g] ",ray_location.x(),ray_location.y(), ray_location.z());
-    printf( "inv_dir [%g,%g,%g] ",inv_ray_direction.x(),inv_ray_direction.y(), inv_ray_direction.z());
-    printf( "disMin %g \n",disMin );
+if( isDbgCell( origin )){
+   printf( "            cur [%d,%d,%d] prev [%d,%d,%d] ", cur.x(), cur.y(), cur.z(), prevCell.x(), prevCell.y(), prevCell.z());
+   printf( " dir %d ", dir );
+   printf( "tMax [%g,%g,%g] ",tMax.x(),tMax.y(), tMax.z());
+   printf( "rayLoc [%g,%g,%g] ",ray_location.x(),ray_location.y(), ray_location.z());
+   printf( "inv_dir [%g,%g,%g] ",inv_ray_direction.x(),inv_ray_direction.y(), inv_ray_direction.z());
+   printf( "disMin %g \n",disMin );
 
-    printf( "            abskg[prev] %g  \t sigmaT4OverPi[prev]: %g \n",abskg[prevCell],  sigmaT4OverPi[prevCell]);
-    printf( "            abskg[cur]  %g  \t sigmaT4OverPi[cur]:  %g  \t  cellType: %i \n",abskg[cur], sigmaT4OverPi[cur], celltype[cur]);
+   printf( "            abskg[prev] %g  \t sigmaT4OverPi[prev]: %g \n",abskg[prevCell],  sigmaT4OverPi[prevCell]);
+   printf( "            abskg[cur]  %g  \t sigmaT4OverPi[cur]:  %g  \t  cellType: %i\n",abskg[cur], sigmaT4OverPi[cur], celltype[cur]);
 }
 #endif
 /*===========TESTING==========`*/
 //cout << "cur " << cur << " face " << face << " tmax " << tMax << " rayLoc " << ray_location <<
 //        " inv_dir: " << inv_ray_direction << " disMin: " << disMin << endl;
 
-       in_domain = (celltype[cur]==-1);  //cellType of -1 is flow
+      in_domain = (celltype[cur] == d_flowCell);
 
 
-       optical_thickness += Dx.x() * abskg_prev*disMin; // as long as tDeltaY,Z tMax.y(),Z and ray_location[1],[2]..
-       // were adjusted by DyDx  or DzDx, this line is now correct for noncubic domains.
+      optical_thickness += Dx.x() * abskg_prev*disMin; // as long as tDeltaY,Z tMax.y(),Z and ray_location[1],[2]..
+      // were adjusted by DyDx  or DzDx, this line is now correct for noncubic domains.
 
-       nRaySteps++;
+      nRaySteps++;
 
-       //Eqn 3-15(see below reference) while
-       //Third term inside the parentheses is accounted for in Inet. Chi is accounted for in Inet calc.
-       double expOpticalThick = exp(-optical_thickness);
+      //Eqn 3-15(see below reference) while
+      //Third term inside the parentheses is accounted for in Inet. Chi is accounted for in Inet calc.
+      double expOpticalThick = exp(-optical_thickness);
 
-       sumI += sigmaT4OverPi_prev * ( expOpticalThick_prev - expOpticalThick ) * fs;
+      sumI += sigmaT4OverPi_prev * ( expOpticalThick_prev - expOpticalThick ) * fs;
 
-       expOpticalThick_prev = expOpticalThick;
+      expOpticalThick_prev = expOpticalThick;
 
 #ifdef RAY_SCATTER
-       curLength += disMin * Dx.x();
+      curLength += disMin * Dx.x();
 
-       if (curLength > scatLength && in_domain){
+      if (curLength > scatLength && in_domain){
 
-         // get new scatLength for each scattering event
-         scatLength = -log(mTwister.randDblExc() ) / scatCoeff;
+        // get new scatLength for each scattering event
+        scatLength = -log(mTwister.randDblExc() ) / scatCoeff;
 
-         ray_direction     =  findRayDirection( mTwister, d_isSeedRandom, cur );
-         inv_ray_direction = Vector(1.0)/ray_direction;
+        ray_direction     =  findRayDirection( mTwister, d_isSeedRandom, cur );
+        inv_ray_direction = Vector(1.0)/ray_direction;
 
-         // get new step and sign
-         int stepOld = step[face];
-         findStepSize( step, sign, inv_ray_direction);
+        // get new step and sign
+        int stepOld = step[dir];
+        findStepSize( step, sign, inv_ray_direction);
 
-         // if sign[face] changes sign, put ray back into prevCell (back scattering)
-         // a sign change only occurs when the product of old and new is negative
-         if( step[face] * stepOld < 0 ){
-           cur = prevCell;
-         }
+        // if sign[dir] changes sign, put ray back into prevCell (back scattering)
+        // a sign change only occurs when the product of old and new is negative
+        if( step[dir] * stepOld < 0 ){
+          cur = prevCell;
+        }
 
-         // get new tMax (mixing bools, ints and doubles)
-         tMax.x( ( cur[0] + sign[0]                - ray_location[0]) * inv_ray_direction[0] );
-         tMax.y( ( cur[1] + sign[1] * D_DxRatio[1] - ray_location[1]) * inv_ray_direction[1] );
-         tMax.z( ( cur[2] + sign[2] * D_DxRatio[2] - ray_location[2]) * inv_ray_direction[2] );
+        // get new tMax (mixing bools, ints and doubles)
+        tMax.x( ( cur[0] + sign[0]                - ray_location[0]) * inv_ray_direction[0] );
+        tMax.y( ( cur[1] + sign[1] * D_DxRatio[1] - ray_location[1]) * inv_ray_direction[1] );
+        tMax.z( ( cur[2] + sign[2] * D_DxRatio[2] - ray_location[2]) * inv_ray_direction[2] );
 
-         // Length of t to traverse one cell
-         tDelta    = Abs(inv_ray_direction) * D_DxRatio;
-         tMax_prev = 0;
-         curLength = 0;  // allow for multiple scattering events per ray
+        // Length of t to traverse one cell
+        tDelta    = Abs(inv_ray_direction) * D_DxRatio;
+        tMax_prev = 0;
+        curLength = 0;  // allow for multiple scattering events per ray
 /*`==========TESTING==========*/
 #if DEBUG == 3
-  printf( "%i, %i, %i, tmax: %g, %g, %g  tDelta: %g, %g, %g \n", cur.x(), cur.y(), cur.z(), tMax.x(), tMax.y(), tMax.z(), tDelta.x(), tDelta.y() , tDelta.z());
+ printf( "%i, %i, %i, tmax: %g, %g, %g  tDelta: %g, %g, %g \n", cur.x(), cur.y(), cur.z(), tMax.x(), tMax.y(), tMax.z(), tDelta.x(), tDelta.y() , tDelta.z());
 #endif
 /*===========TESTING==========`*/
-         //if(_benchmark == 4 || _benchmark ==5) scatLength = 1e16; // only for Siegel Benchmark4 benchmark5. Only allows 1 scatter event.
-       }
+      }
 #endif
 
-     } //end domain while loop.  ++++++++++++++
+    } //end domain while loop.  ++++++++++++++
 
-     T wallEmissivity = abskg[cur];
+    T wallEmissivity = abskg[cur];
 
-     if (wallEmissivity > 1.0){       // Ensure wall emissivity doesn't exceed one.
-       wallEmissivity = 1.0;
-     }
+    if (wallEmissivity > 1.0){       // Ensure wall emissivity doesn't exceed one.
+      wallEmissivity = 1.0;
+    }
 
-     intensity = exp(-optical_thickness);
+    intensity = exp(-optical_thickness);
 
-     sumI += wallEmissivity * sigmaT4OverPi[cur] * intensity;
+    sumI += wallEmissivity * sigmaT4OverPi[cur] * intensity;
 
-     intensity = intensity * fs;
+    intensity = intensity * fs;
 
-     // when a ray reaches the end of the domain, we force it to terminate.
-     if(!d_allowReflect) intensity = 0;
+    // when a ray reaches the end of the domain, we force it to terminate.
+    if(!d_allowReflect) intensity = 0;
 
 /*`==========TESTING==========*/
 #if DEBUG == 1
-if(origin.x() == 0 && origin.y() == 0 && origin.z() ==0 ){
-    printf( "            cur [%d,%d,%d] intensity: %g expOptThick: %g, fs: %g allowReflect: %i\n",
-           cur.x(), cur.y(), cur.z(), intensity,  exp(-optical_thickness), fs, d_allowReflect );
+if( isDbgCell( origin)  ){
+   printf( "            cur [%d,%d,%d] intensity: %g expOptThick: %g, fs: %g allowReflect: %i\n",
+          cur.x(), cur.y(), cur.z(), intensity,  exp(-optical_thickness), fs, d_allowReflect );
 
 }
 #endif
 /*===========TESTING==========`*/
-     //__________________________________
-     //  Reflections
-     if ( (intensity > d_threshold) && d_allowReflect){
-       reflect( fs, cur, prevCell, abskg[cur], in_domain, step[face], sign[face], ray_direction[face]);
-       ++nReflect;
-     }
-   }  // threshold while loop.
+    //__________________________________
+    //  Reflections
+    if ( (intensity > d_threshold) && d_allowReflect){
+      reflect( fs, cur, prevCell, abskg[cur], in_domain, step[dir], sign[dir], ray_direction[dir]);
+      ++nReflect;
+    }
+  }  // threshold while loop.
 } // end of updateSumI function
 
 
@@ -663,6 +672,25 @@ RMCRTCommon::doCarryForward( const int radCalc_freq ){
   bool test = (timestep%radCalc_freq != 0 && timestep != 1);
 
   return test;
+}
+
+//______________________________________________________________________
+//
+//______________________________________________________________________
+bool
+RMCRTCommon::isDbgCell( const IntVector me)
+{
+
+  if(me == IntVector(10,10,10) ){
+    return true;
+  }
+  for( unsigned int i = 0; i<d_dbgCells.size(); i++) {
+    if( me == d_dbgCells[i]) {
+      return true;
+
+    }
+  }
+  return false;
 }
